CPO-8 - rename of integration test class to match master branch
